fix: correctly pass empty strings (#74)
* Fix #73.

Empty string is encoded as empty buffer in Deno. And as of
right now (Deno 1.29.1), ffi layer converts it to NULL pointer,
which causes sqlite3_bind_text to bind the NULL value instead
of an empty string. As a workaround let's use a special
non-empty buffer, but specify zero length.

* Avoid creating a new database in a test.

* Fix formatting.